Create WES237A Lab 1 Question 3
WES 237A Lab1: Write code in the section ‘Lab Work’ to use two pins (0 and 1) for send and two pins (2
and 3) for receive. You should be able to send 2 bits (0~3) over GPIO. You'll need to
hardwire from the send pins to the receive pins.
a. Start the code by copying ‘cells’ 1 and 2 from the beginning of the notebook into
the ‘Lab Work’ section.
b. Then begin editing the %%microblaze cell.